fix/adjustments in posts services
