+++ src/components/TimeLine/UserPostForm.component.jsx
@@ -1,21 +1,100 @@
+import { AuthContext } from "../context/auth.context";
+import { useCallback, useEffect, useContext, useState} from "react";
 import  styled  from "styled-components";
-import { useCallback } from "react";
+import axios from "axios";
 
 
-export default function PostForm({url, setUrl, publishing, description, setDescription, handlePost}) {
+export default function PostForm() {
 
+  const { token } = useContext(AuthContext);
+  const [url, setUrl] = useState("");
+  const [description, setDescription] = useState("");
+  const [error, setError] = useState(false);
+  const [emptyPage, setEmptyPage] = useState(false);
+  const [publishing, setPublishing] = useState(false);
+  const [refresh, setRefresh] = useState(true);
+  const [posts, setPosts] = useState([]);
+  const config = { headers: { Authorization: `Bearer ${token}` } };
   const authorImage =
       "https://encrypted-tbn3.gstatic.com/images?q=tbn:ANd9GcTfeiK25FERClFs4W7YW5U9uN3EgWX1istoqeFeN_IPFLBGOvaC";
-  
 
-    const buttonPressed = useCallback(
-      (event) => {
-        if (event.key === "Enter") {
-          handlePost();
-        }
-      },
-      [handlePost]
-  );
+
+//CARREGAR POSTS PUBLICADOS (20 mais recentes);
+useEffect(() => {
+  axios
+    .get(`${process.env.REACT_APP_API_URL}/posts`, config)
+    .then((res) => {
+
+      if (Array.isArray(res.data.results)) { 
+        const sortPosts = res.data.results.sort((i, j) => j.id - i.id);
+        const renderPosts = sortPosts.slice(0, 20);
+        setPosts(renderPosts);
+        setEmptyPage(renderPosts.length === 0);
+        setRefresh(false);
+      } 
+      else {
+        console.error(res.data);
+        setError(true);
+        setRefresh(false);
+        alert(
+          "An error occured while trying to fetch the posts, please refresh the page."
+        );
+      }
+    })
+    .catch((err) => {
+      console.error(err);
+      setError(true);
+      setRefresh(false);
+      alert(
+        "An error occured while trying to fetch the posts, please refresh the page."
+      );
+    });
+  // eslint-disable-next-line
+}, []);
+
+
+//PUBLICAR POSTS;
+  const handlePost = useCallback(async () => {
+  if (url === "") {
+    alert("Please, enter the url of your post!");
+    return;
+  }
+  setPublishing(true);
+
+  try {
+      await axios.post(
+          `${process.env.REACT_APP_API_URL}/posts`,
+          {
+              url: url,
+              description: description,
+          },
+          config
+      );
+      setUrl("");
+      setDescription("");
+
+      
+    //POSTS ATUALIZADOS;
+      const updatedPosts = await axios.get(
+          `${process.env.REACT_APP_API_URL}/posts`, config);
+
+      const sortPosts = updatedPosts.data.sort((i, j) => j.id - i.id);
+      const renderPosts = sortPosts.slice(0, 20);
+
+      setPosts(renderPosts);
+      setEmptyPage(renderPosts.length === 0);
+      setUrl("");
+      setDescription("");
+  }
+  catch (err) {
+      console.error(err);
+      alert("There was an error publishing your link.");
+  } finally {
+      setPublishing(false);
+  }
+// eslint-disable-next-line
+}, [handlePost]);   
+    
 
   return (
       <PostContainer data-test="publish-box">
@@ -48,7 +127,7 @@ export default function PostForm({url, setUrl, publishing, description, setDescr
 
             <button
                 data-test="publish-btn"
-                onClick={buttonPressed}
+                onClick={handlePost}
                 disabled={publishing}
               >
                 {publishing ? "Publishing..." : "Publish"}
@@ -56,8 +135,9 @@ export default function PostForm({url, setUrl, publishing, description, setDescr
           </FormContainer>
         </PublishBox>
       </PostContainer>
-  )
-}
+    )
+  }
+
 
 const PostContainer = styled.div`
   font-family: "Lato";
@@ -115,16 +195,16 @@ const FormContainer = styled.form`
     font-size: 20px;
     background-color: #efefef;
     height: 30px;
+       
+    &:disabled {
+      background-color: #d3d3d3;
+      cursor: not-allowed;
+    }
   }
   input::placeholder {
     font-family: "Lato";
     font-weight: 300;
     color: #949494;
-    
-    &:disabled {
-      background-color: #d3d3d3;
-      cursor: not-allowed;
-    }
   }
   :focus {
     outline: none;
